refactor(grouping): collapse grouper strategy into a single function
The MutationGrouper interface seam wasn't paying its way: NoOp was a
trivial inline behaviour, DSatur was the only real implementation, and
no third strategy is on the roadmap.

Changes:
- mutationGrouper.ts now exports groupMutations() — the DSATUR algorithm
  as a free function. Drops the MutationGrouper interface, GrouperInput
  interface, MutableGroup helper, and NoOpMutationGrouper class.
- planGroups in MutationTestingService inlines the singleton-grouping
  behaviour for the default-off case (no conflict graph built).
- Deletes src/service/groupers/ and test/unit/service/groupers/. Tests
  for groupMutations + assertGroupingInvariants live in the existing
  mutationGrouper.test.ts with inlined fixture helpers.

Net: 168 fewer lines, 4 fewer interfaces (MutationGrouper, GrouperInput,
MutableGroup; NoOp class also gone). Only MutationGroup remains.
Behaviour preserved: 1551 unit tests pass at 100% branch coverage.
